# Conflicts:
#	cmp-navigation/src/commonMain/kotlin/cmp/navigation/navigation/FeatureNavHost.kt

+++ core/model/src/commonMain/kotlin/com/mifos/core/model/objects/organisations/LoanProducts.kt
@@ -160,7 +160,7 @@ data class LoanProducts(
     var maxTrancheCount: Int? = null,
 
     @SerialName("principalThresholdForLastInstallment")
-    var principalThresholdForLastInstallment: Int? = null,
+    var principalThresholdForLastInstallment: Double? = null,
 
     @SerialName("holdGuaranteeFunds")
     var holdGuaranteeFunds: Boolean? = null,
